Finish: Code Refactoring and Getting Network Interface Information.
- Code Refactoring: Edit Data Type, equal test logic, function name, and so on.
- Getting network interface information (Interface name, IPv4 Address, Speed, Error Packet, Drop packet)

+++ cpu_info_struct.h
@@ -12,8 +12,8 @@ typedef struct CPU_Result {
 } CPU_Result;
 
 typedef struct CPU_Usage {
-    long long totalJiff;
-    long long idleJiff;
+    unsigned long totalJiff;
+    unsigned long idleJiff;
 } CPU_Usage;
 
 typedef struct CPU_Info {
